Re-organize home care package factories
