+++ LBH.AdultSocialCare.Api/V1/Controllers/HomeCare/HomeCarePackageController.cs
@@ -1,22 +1,19 @@
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Threading.Tasks;
 using LBH.AdultSocialCare.Api.V1.Boundary.HomeCareApprovalHistoryBoundary.Response;
 using LBH.AdultSocialCare.Api.V1.Boundary.Request.HomeCare;
 using LBH.AdultSocialCare.Api.V1.Boundary.Response;
-using LBH.AdultSocialCare.Api.V1.Domain;
-using LBH.AdultSocialCare.Api.V1.Domain.HomeCare;
 using LBH.AdultSocialCare.Api.V1.Factories;
 using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.HomeCare;
 using LBH.AdultSocialCare.Api.V1.UseCase.HomeCareApprovalHistoryUseCase.Interfaces;
 using LBH.AdultSocialCare.Api.V1.UseCase.Interfaces;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Threading.Tasks;
 
 namespace LBH.AdultSocialCare.Api.V1.Controllers.HomeCare
 {
-
     [Route("api/v1/[controller]")]
     [Produces("application/json")]
     [ApiController]
@@ -24,7 +21,6 @@ namespace LBH.AdultSocialCare.Api.V1.Controllers.HomeCare
     [ApiVersion("1.0")]
     public class HomeCarePackageController : BaseController
     {
-
         private readonly IUpsertHomeCarePackageUseCase _upsertHomeCarePackageUseCase;
         private readonly IChangeStatusHomeCarePackageUseCase _updateHomeCarePackageUseCase;
         private readonly IGetAllHomeCarePackageUseCase _getAllHomeCarePackageUseCase;
@@ -56,10 +52,11 @@ namespace LBH.AdultSocialCare.Api.V1.Controllers.HomeCare
         {
             try
             {
-                HomeCarePackageResponse homeCarePackageResponse =
-                    HomeCarePackageFactory.ToResponse(await _updateHomeCarePackageUseCase
-                        .UpdateAsync(homeCarePackageId, statusId)
-                        .ConfigureAwait(false));
+                var res = await _updateHomeCarePackageUseCase
+                    .UpdateAsync(homeCarePackageId, statusId)
+                    .ConfigureAwait(false);
+                var homeCarePackageResponse =
+                    res.ToResponse();
                 if (homeCarePackageResponse == null) return NotFound();
                 return Ok(homeCarePackageResponse);
             }
@@ -84,12 +81,13 @@ namespace LBH.AdultSocialCare.Api.V1.Controllers.HomeCare
         {
             try
             {
-                HomeCarePackageDomain homeCarePackageDomain = HomeCarePackageFactory.ToDomain(homeCarePackageRequest);
+                var homeCarePackageDomain = homeCarePackageRequest.ToDomain();
 
-                HomeCarePackageResponse homeCarePackageResponse =
-                    HomeCarePackageFactory.ToResponse(await _upsertHomeCarePackageUseCase
-                        .ExecuteAsync(homeCarePackageDomain)
-                        .ConfigureAwait(false));
+                var res = await _upsertHomeCarePackageUseCase
+                    .ExecuteAsync(homeCarePackageDomain)
+                    .ConfigureAwait(false);
+
+                var homeCarePackageResponse = res.ToResponse();
 
                 if (homeCarePackageResponse == null)
                 {
@@ -121,7 +119,7 @@ namespace LBH.AdultSocialCare.Api.V1.Controllers.HomeCare
         {
             try
             {
-                IList<HomeCarePackage> result = await _getAllHomeCarePackageUseCase.GetAllAsync().ConfigureAwait(false);
+                var result = await _getAllHomeCarePackageUseCase.GetAllAsync().ConfigureAwait(false);
                 return Ok(result);
             }
             catch (FormatException ex)
@@ -141,5 +139,4 @@ namespace LBH.AdultSocialCare.Api.V1.Controllers.HomeCare
             return Ok(result);
         }
     }
-
 }
