+++ LBH.AdultSocialCare.Api/V1/Domain/DayCarePackageReclaimDomains/DayCarePackageClaimCreationDomain.cs
@@ -1,7 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace LBH.AdultSocialCare.Api.V1.Domain.DayCarePackageReclaimDomains
 {
