+++ LBH.AdultSocialCare.Api/V1/UseCase/HomeCare/ChangeStatusHomeCarePackageUseCase.cs
@@ -1,16 +1,16 @@
-using LBH.AdultSocialCare.Api.V1.Domain;
+using LBH.AdultSocialCare.Api.V1.Domain.HomeCare;
 using LBH.AdultSocialCare.Api.V1.Factories;
 using LBH.AdultSocialCare.Api.V1.Gateways.Interfaces;
 using LBH.AdultSocialCare.Api.V1.UseCase.Interfaces;
 using System;
 using System.Threading.Tasks;
-using LBH.AdultSocialCare.Api.V1.Domain.HomeCare;
 
 namespace LBH.AdultSocialCare.Api.V1.UseCase.HomeCare
 {
     public class ChangeStatusHomeCarePackageUseCase : IChangeStatusHomeCarePackageUseCase
     {
         private readonly IHomeCarePackageGateway _gateway;
+
         public ChangeStatusHomeCarePackageUseCase(IHomeCarePackageGateway homeCarePackageGateway)
         {
             _gateway = homeCarePackageGateway;
@@ -19,11 +19,7 @@ namespace LBH.AdultSocialCare.Api.V1.UseCase.HomeCare
         public async Task<HomeCarePackageDomain> UpdateAsync(Guid homeCarePackageId, int statusId)
         {
             var homeCarePackageEntity = await _gateway.ChangeStatusAsync(homeCarePackageId, statusId).ConfigureAwait(false);
-            if (homeCarePackageEntity == null) return null;
-            else
-            {
-                return HomeCarePackageFactory.ToDomain(homeCarePackageEntity);
-            }
+            return homeCarePackageEntity?.ToDomain();
         }
     }
 }
